Typecast changes

+++ src/gui/zipfile.h
@@ -72,7 +72,7 @@ class ZIPFileEntry {
 public:
     bool        can_write = false;
     bool        can_extend = false;
-    uint32_t    file_length = 0;
+    off_t       file_length = 0;
     off_t       file_offset = 0;
     off_t       file_header_offset = 0;
     off_t       position = 0;
